+++ example/lib/main.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:skiff/skiff.dart';
 
-class FetchCharacters implements Query<FetchCharactersResult> {}
+class FetchCharacters implements Request<FetchCharactersResult> {}
 
 class FetchCharactersResult {
   final List<String> characters;
@@ -10,7 +10,7 @@ class FetchCharactersResult {
   FetchCharactersResult(this.characters);
 }
 
-class Authenticate implements Command {
+class Authenticate implements Request<RequestResult> {
   final String username;
   final String password;
 
@@ -23,7 +23,7 @@ void main() async {
   var query = FetchCharacters();
 
   var charactersHandler =
-      FuncHandler<FetchCharacters, FetchCharactersResult>((command) {
+      FuncHandler<FetchCharacters, FetchCharactersResult>((r) {
     return Future.value(FetchCharactersResult([
       'Eleven',
       'Mike Wheeler',
@@ -41,13 +41,12 @@ void main() async {
 
   var command = Authenticate('alexei@starcourtmall.com', 'pa55w0rd');
 
-  var authenticationHandler =
-      FuncHandler<Authenticate, CommandResult>((command) {
-    return Future.value(CommandResult.succeeded());
+  var authenticationHandler = FuncHandler<Authenticate, RequestResult>((r) {
+    return Future.value(RequestResult.succeeded());
   });
 
   mediator.addHandler(authenticationHandler);
 
-  var result = await mediator.dispatch<CommandResult>(command);
+  var result = await mediator.dispatch<RequestResult>(command);
   print('Login ${result.isSuccessful ? 'successful' : 'failed'}');
 }