+++ lib/skiff.dart
@@ -1,18 +1,11 @@
-/// A lightweight, Future-based library for expressing CQRS principles in Dart
-/// projects.
-///
-/// This package allows you to write smaller, more robust and testable code by
-/// turning data access (or anything else you can think of) logic into first-class
-/// objects with sound typing.
 library skiff;
 
-export 'src/commands/command.dart';
-export 'src/commands/command_handler.dart';
-export 'src/commands/command_result.dart';
-
-export 'src/queries/query.dart';
-export 'src/queries/query_handler.dart';
-
-export 'src/handler.dart' show FuncHandler;
-
+export 'src/bus.dart';
+export 'src/event.dart';
+export 'src/exceptions.dart';
+export 'src/handler.dart';
 export 'src/mediator.dart';
+export 'src/registration.dart';
+export 'src/relay.dart';
+export 'src/request.dart';
+export 'src/request_result.dart';