+++ lib/src/mediator.dart
@@ -1,23 +1,45 @@
 import 'dart:async';
 import 'dart:collection';
 
+import 'package:skiff/src/bus.dart';
+import 'package:skiff/src/event.dart';
 import 'package:skiff/src/exceptions.dart';
 import 'package:skiff/src/handler.dart';
+import 'package:skiff/src/registration.dart';
+import 'package:skiff/src/relay.dart';
 import 'package:skiff/src/request.dart';
+import 'package:skiff/src/stream_registration.dart';
 
-/// Promotes loose coupling between caller and callees.
-class Mediator {
+/// Promotes loose coupling between caller and callees by orchestrating the
+/// dispatch of [Request] or [Event] objects.
+class Mediator implements Bus, Relay {
   final Map<Type, Handler> _handlers = {};
+  final StreamController<Event> _events = StreamController.broadcast();
 
-  /// Gets the currently registered handlers for this [Mediator].
-  UnmodifiableMapView<Type, Handler> get handlers =>
+  @override
+  UnmodifiableMapView<Type, Handler<Request, dynamic>> get handlers =>
       UnmodifiableMapView(_handlers);
 
-  /// Adds a handler for the request type [R].
-  ///
-  /// This [handler] will be registered against the type [R] and executed
-  /// whenever a [Request] of that type is dispatched.
+  bool _disposed = false;
+  bool get isDisposed => _disposed;
+
+  bool get _canAdd => !isDisposed;
+
+  void dispose() {
+    _events.close();
+    _disposed = true;
+  }
+
+  @override
+  Registration listen<E extends Event>(Function(E) fn) {
+    _throwBadStateIfDisposed("add listener");
+    return StreamRegistration(
+        _events.stream.where((e) => e.runtimeType == E).cast<E>().listen(fn));
+  }
+
+  @override
   void addHandler<R extends Request>(Handler<R, dynamic> handler) {
+    _throwBadStateIfDisposed("add handler");
     if (_handlers.containsKey(R)) {
       throw AlreadyRegisteredException(R);
     }
@@ -25,8 +47,9 @@ class Mediator {
     _handlers[R] = handler;
   }
 
-  /// Removes a handler for the request type [R].
+  @override
   Handler<Request<dynamic>, dynamic> removeHandler<R extends Request>() {
+    _throwBadStateIfDisposed("remove handler");
     if (R == Request) {
       throw RequestTypeMissingException();
     }
@@ -36,19 +59,27 @@ class Mediator {
     return _handlers.remove(R)!;
   }
 
-  /// Passes the given [request] to a registered handler and returns a response
-  /// of type [Rs].
+  @override
   Future<Rs> dispatch<Rs>(Request<Rs> request) async {
+    _throwBadStateIfDisposed("dispatch");
     _throwIfHandlerMissing(request.runtimeType);
 
     return await _handlers[request.runtimeType]?.execute(request);
   }
 
-  void _throwIfHandlerMissing(Type requestType) {
-    if (_handlers.containsKey(requestType)) {
-      return;
-    }
+  @override
+  void broadcast<E extends Event>(E event) {
+    _throwBadStateIfDisposed("broadcast");
+    _events.add(event);
+  }
 
+  void _throwBadStateIfDisposed(String action) {
+    if (_canAdd) return;
+    throw StateError("Cannot $action after disposal");
+  }
+
+  void _throwIfHandlerMissing(Type requestType) {
+    if (_handlers.containsKey(requestType)) return;
     throw MissingHandlerException(requestType);
   }
 }