+++ lib/src/registration.dart
@@ -1,0 +1,4 @@
+/// Supplies behaviour for safe disposal of listeners.
+abstract class Registration {
+  void dispose();
+}