+++ test/helpers/mock_event_callback.dart
@@ -1,0 +1,5 @@
+import 'package:mocktail/mocktail.dart';
+
+class MockEventCallback extends Mock {
+  dynamic call(dynamic e);
+}